Refactor code and improve code quality
Previously, the code was very messy and had inconsistent
levels of abstractions throughout.

Created different packages and moved classes into their
respective packages depending on their function.

This provides an easier way to read the code and check for
bugs and logic errors.
