+++ src/main/java/duke/Duke.java
@@ -1,5 +1,11 @@
 package duke;
 
+import duke.exceptions.DukeException;
+import duke.gui.Main;
+import duke.logic.Executor;
+import duke.logic.Storage;
+import duke.logic.Ui;
+import duke.tasks.TaskList;
 import javafx.application.Application;
 
 /**
