+++ src/main/java/duke/exceptions/DukeException.java
@@ -1,9 +1,9 @@
-package duke;
+package duke.exceptions;
 
 /**
  * Special exceptions encountered by the chatbot.
  */
-class DukeException extends Exception {
+public class DukeException extends Exception {
     /**
      * Constructor for the DukeException class.
      * @param errorMessage Message about the error.
