+++ src/main/java/duke/exceptions/EmptyDescription.java
@@ -1,15 +1,15 @@
-package duke;
+package duke.exceptions;
 
 /**
  * Exception when no description is found.
  */
-class EmptyDescription extends DukeException {
+public class EmptyDescription extends DukeException {
     /**
      * Constructor for the EmptyDescription class.
      */
     public EmptyDescription() {
         super("-------------------------------\n"
-                + "☹ OOPS!!! The description of a todo cannot be empty.\n"
+                + "OOPS!!! The description of a todo cannot be empty.\n"
                 + "-------------------------------\n");
     }
 }
