+++ src/test/java/duke/TodoTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.tasks.Todo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
